fixing seed test and seed file setup
