+++ script/seed.js
@@ -1,58 +1,93 @@
-'use strict'
+const db = require(`../server/db`)
+const { Product, User, Review } = require(`../server/db/models`)
 
-const db = require('../server/db')
-const {User} = require('../server/db/models')
-
-/**
- * Welcome to the seed file! This seed file uses a newer language feature called...
- *
- *                  -=-= ASYNC...AWAIT -=-=
- *
- * Async-await is a joy to use! Read more about it in the MDN docs:
- *
- * https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/async_function
- *
- * Now that you've got the main idea, check it out in practice below!
- */
+const grasses = [
+  {
+    name: `Grasp Grass`,
+    price: 2.5,
+    imageUrl:
+      `https://previews.123rf.com/images/nataliafedorenko/nataliafedorenko0905/nataliafedorenko090500053/4867643-a-male-hand-with-its-fingers-picking-up-the-grass-.jpg`,
+    description: `Grass that grabs you`
+  },
+  {
+    name: `Dead grass`,
+    price: 1,
+    imageUrl:
+      `https://forums.unrealengine.com/filedata/fetch?id=1211751&d=1487743042`,
+    description:
+      `Tired of smelling the dew. Well you can buy dead grass now and just smell burnt plants`
+  },
+  {
+    name: `Dead Grass on fire`,
+    price: 0.5,
+    imageUrl:
+      `http://mediad.publicbroadcasting.net/p/wnmu/files/styles/medium/public/201805/grass_fire.jpg`,
+    description:
+      `Ok if you got tired of smelling grass and dead grass, how about now you just smell grass thats on fire. Must be 18 or older to purchase.`
+  }
+]
+const users = [
+  {
+    name: `Kenneth Lai`,
+    street: `hood`,
+    city: `Rack City`,
+    state: `Rack State`,
+    zipcode: 51134,
+    email: `knthslai@gmail.com`,
+  }, {
+    name: `Ali Aftab`,
+    street: `hooder`,
+    city: `Hooder City`,
+    state: `Detroit`,
+    zipcode: 58008,
+    email: `itisaftab@geocities.com`,
+  }
+]
+const reviews = [
+  {
+    title: `G is for Grass`,
+    body: `H is for happyness. - Will Smith`,
+    userId: 1,
+    productId: 2
+  }, {
+    title: `My house is on fire ...`,
+    body: `I should not have boughten "Dead Grass on fire"...`,
+    userId: 2,
+    productId: 3,
+  }
+]
+// const seed = () => Promise.all(users.map(user => User.create(user))).then(() => Promise.all(grasses.map(grass => Product.create(grass)))).then(() => Promise.all(reviews.map(review => Review.create(review))))
 
 async function seed() {
-  await db.sync({force: true})
-  console.log('db synced!')
-  // Whoa! Because we `await` the promise that db.sync returns, the next line will not be
-  // executed until that promise resolves!
-  const users = await Promise.all([
-    User.create({email: 'cody@email.com', password: '123'}),
-    User.create({email: 'murphy@email.com', password: '123'})
-  ])
-  // Wowzers! We can even `await` on the right-hand side of the assignment operator
-  // and store the result that the promise resolves to in a variable! This is nice!
-  console.log(`seeded ${users.length} users`)
+  await db.sync({ force: true })
+  console.log(`db synced!`)
+
+  const usersP = await Promise.all(users.map(user => User.create(user)))
+  const grassesP = await Promise.all(grasses.map(grass => Product.create(grass)))
+  const reviewsP = await Promise.all(reviews.map(review => Review.create(review)))
   console.log(`seeded successfully`)
 }
 
-// We've separated the `seed` function from the `runSeed` function.
-// This way we can isolate the error handling and exit trapping.
-// The `seed` function is concerned only with modifying the database.
-async function runSeed() {
-  console.log('seeding...')
-  try {
-    await seed()
-  } catch (err) {
-    console.error(err)
-    process.exitCode = 1
-  } finally {
-    console.log('closing db connection')
-    await db.close()
-    console.log('db connection closed')
-  }
+const main = () => {
+  console.log(`Syncing db...`)
+  db
+    .sync({ force: true })
+    .then(() => {
+      console.log(`Seeding databse...`)
+      return seed()
+    })
+    .catch(err => {
+      console.log(`Error while seeding`)
+      console.log(err.stack)
+    })
+    .then(() => {
+      console.log(`Seeded successfully`)
+      db.close()
+      return null
+    })
 }
-
-// Execute the `seed` function, IF we ran this module directly (`node seed`).
-// `Async` functions always return a promise, so we can use `catch` to handle
-// any errors that might occur inside of `seed`.
 if (module === require.main) {
-  runSeed()
+  main()
 }
 
-// we export the seed function for testing purposes (see `./seed.spec.js`)
 module.exports = seed
