+++ server/api/users.js
@@ -4,12 +4,7 @@ module.exports = router
 
 router.get(`/`, async (req, res, next) => {
   try {
-    const users = await User.findAll({
-      // explicitly select only the id and email fields - even though
-      // users' passwords are encrypted, it won't help if we just
-      // send everything to anyone who asks!
-      attributes: [`id`, `email`]
-    })
+    const users = await User.findAll()
     res.json(users)
   } catch (err) {
     next(err)
